+++ android/app/src/main/res/values/styles.xml
@@ -12,4 +12,7 @@
   <style name="Theme.App.SplashScreen" parent="AppTheme">
     <item name="android:windowBackground">@drawable/splashscreen_logo</item>
   </style>
+  <style name="DatePickerTheme" parent="DatePickerBaseTheme">
+    <item name="android:textSize">18dp</item>
+  </style>
 </resources>
