+++ package.json
@@ -13,6 +13,8 @@
     "@react-native-picker/picker": "^2.11.0",
     "axios": "^1.7.9",
     "expo": "~52.0.28",
+    "expo-audio": "~0.3.4",
+    "expo-av": "~15.0.2",
     "expo-constants": "~17.0.5",
     "expo-linking": "~7.0.5",
     "expo-router": "~4.0.17",
@@ -20,14 +22,13 @@
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.6",
+    "react-native-date-picker": "^5.0.10",
     "react-native-dotenv": "^3.4.11",
     "react-native-floating-bubble-plugin": "^1.0.71",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-webview": "13.12.5",
-    "react-native-youtube-iframe": "^2.3.0",
-    "expo-audio": "~0.3.4",
-    "expo-av": "~15.0.2"
+    "react-native-youtube-iframe": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
